refactor(openehr-rm): the RM typed validation dispatch moves beside validate::fast
The wire_validate justification expired with the foundation rewrite
(#1702): openehr-rm carries the emitted manual serde impls, so the
_type → run::<T> dispatch table, run/run_shallow and their decode now
live in openehr_rm::validate::typed_dispatch (dispatch_typed returns
whether the type was handled). openehr-its keeps only what genuinely
spans the five spec crates — the generated structural fallthrough — plus
the thin wire entry (undeclared-key refusal → dispatch → structural →
the LOCATABLE node-id closeout, ordering unchanged). Decode reproduces
from_canonical_value byte-for-byte at the message level; the
dispatch-lockstep gate repoints at the new home with an unchanged
snapshot; the stale topology comments in rm_validation.rs and both
crate memories corrected.

Closes #1823
