+++ crates/openehr-rm/src/validate/fast.rs
@@ -1,12 +1,14 @@
 //! The allocation-free fast path of the RM class-invariant check (hand-written),
-//! exposed via [`super::try_fast_validate`]. The wire-boundary two-tier entry
-//! point that calls it then falls back to the typed dispatch lives in
+//! exposed via [`super::try_fast_validate`]. The authoritative tier it declines
+//! to is the sibling [`super::typed_dispatch`]; the wire-boundary entry point
+//! that runs the two in order lives in
 //! `openehr_its::wire_validate::validate_rm_value`.
 //!
 //! No openEHR spec governs this module — it is our own performance design; the
 //! *semantics* it realizes are exactly those of the typed dispatch in
-//! [`super`] (the RM class invariants of the `*_impl.rs` siblings plus the
-//! structural type-conformance rejection of a typed deserialize).
+//! [`super::typed_dispatch`] (the RM class invariants of the `*_impl.rs`
+//! siblings plus the structural type-conformance rejection of a typed
+//! deserialize).
 //!
 //! # Design: vouch-or-fall-back
 //!
@@ -56,7 +58,7 @@
 //!   [`fast_spec`], the `Interval` default-able bound flags) → fall back.
 //!
 //! **Shallow mode** mirrors the typed dispatcher's `prune_child_nodes` (in
-//! `openehr_its::wire_validate`) for the structural container classes the typed
+//! [`super::typed_dispatch`]) for the structural container classes the typed
 //! path checks via `run_shallow`: a child
 //! *collection* is vouched without descending iff it is empty or contains at
 //! least one object (exactly the arrays the prune empties before the typed
@@ -88,7 +90,7 @@ pub(super) fn try_validate(ty: &str, value: &Value, out: &mut Vec<InvariantViola
     // Dispatch mode: which classes have a fast invariant evaluator, and
     // whether the typed path would deserialize them shallowly (`run_shallow`)
     // or in full (`run`). Must stay in lockstep with the typed dispatch table
-    // in `openehr_its::wire_validate` (`validate_rm_value_typed`).
+    // in `super::typed_dispatch` (`dispatch_typed`).
     let shallow = match ty {
         // `run_shallow` classes (structural containers, scalar-only invariants).
         "CLUSTER" | "POINT_EVENT" | "INTERVAL_EVENT" | "COMPOSITION" | "EVENT_CONTEXT"
